Introduce advice on recharging skills
Currently, a number of iconic creatures have skills which recharge randomly. As an example, an Adult Red Dragon has an action: "Fire Breath (Recharge 5/6)". This system of random recharging is not interactive for players. This proposal introduces more clear rules on how to handle recharging skills in a deterministic way, along with hinting to players that such an attack is incoming.